requirements.txt
Flask
twilio
websockets
google-cloud-speech
google-cloud-texttospeech
openai
python-dotenv